functions/orchestrator/requirements.txt
functions-framework==3.*
google-auth==2.3.3
requests==2.22.0